Move SpacingComponent into the styled folder
