+++ cardigan/stories/global/spacing/spacing.stories.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent } from 'react';
 import styled from 'styled-components';
 import Table from '@weco/common/views/components/Table/Table';
 import SpacingSection from '@weco/common/views/components/styled/SpacingSection';
-import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
+import SpacingComponent from '@weco/common/views/components/styled/SpacingComponent';
 import { PaletteColor, themeValues } from '@weco/common/views/themes/config';
 
 const { spaceAtBreakpoints } = themeValues;
