+++ catalogue/webapp/components/WorkDetailsProperty/WorkDetailsProperty.tsx
@@ -1,5 +1,5 @@
 import { font } from '@weco/common/utils/classnames';
-import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
+import SpacingComponent from '@weco/common/views/components/styled/SpacingComponent';
 import Space from '@weco/common/views/components/styled/Space';
 import ConditionalWrapper from '@weco/common/views/components/ConditionalWrapper/ConditionalWrapper';
 import { FunctionComponent, PropsWithChildren } from 'react';
