+++ common/views/components/styled/SpacingComponent.tsx
