+++ content/webapp/pages/pages/[pageId].tsx
@@ -19,7 +19,7 @@ import {
   sectionLevelPages,
 } from '@weco/common/data/hardcoded-ids';
 import SpacingSection from '@weco/common/views/components/styled/SpacingSection';
-import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
+import SpacingComponent from '@weco/common/views/components/styled/SpacingComponent';
 import SectionHeader from '@weco/common/views/components/SectionHeader/SectionHeader';
 import { PageFormatIds } from '@weco/common/data/content-format-ids';
 import { links } from '@weco/common/views/components/Header/Header';
